오류수정
categoryId 삭제 업데이트
