+++ src/main/java/com/library/repository/BookRepositoryCustom.java
@@ -2,6 +2,7 @@ package com.library.repository;
 
 import com.library.dto.BookListDto;
 import com.library.dto.BookSearchDto;
+import com.library.dto.MainBookDto;
 import com.library.entity.Book;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
